Fix issue #31: rim status is checking untouched modules
When there's a module with .riminfo and there's a change
in a parallel folder having the module's path as prefix
(e.g. module platform/bspMcu and changed folder
platform/bspMcuWatchdog) then rim status will check
the unchanged module. This might lead to confusion in
case this module was dirty before. Fix this problem.